+++ src/components/example/SimpleProp.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-defineProps<{ msg: string }>();
+const props = defineProps<{ msg: string }>();
 </script>
 
 <template>
-  <h1 class="text-4xl font-bold text-white">{{ msg }}</h1>
+  <h1 class="text-4xl font-bold text-white">{{ props.msg }}</h1>
 </template>